// full/src/global-types.ts
// Auto-generated global type definitions for ikas code components.
// This file is regenerated automatically — do not edit manually.

/** Enum type: s3OODUmZpJ */
export type AspectRatio = "Square" | "Portrait" | "Landscape" | "Video";
/** Enum type: Tkt3wfuJ5F */
export type NumberOfColumns = "One" | "Two" | "Three" | "Four" | "Five";
/** Enum type: GrylMqHxui */
export type ObjectFit = "Fill" | "Cover" | "Contain";
/** Enum type: 08UWYbbs0P */
export type VerticalAlignment = "Top" | "Middle" | "Bottom";